+++ src/core/RowWrapper.cpp
@@ -17,7 +17,7 @@ RowWrapper::handle_type RowWrapper::create(const FieldRow row)
 
 bool RowWrapper::is_solved() const
 {
-    for (int x = 0; x < 9; x++) {
+    for (size_t x = 0; x < 9; x++) {
         if (!(data[x]->is_set())) {
             return false;
         }
@@ -28,7 +28,7 @@ bool RowWrapper::is_solved() const
 
 bool RowWrapper::contains(const Value val) const
 {
-    for (int x = 0; x < 9; x++) {
+    for (size_t x = 0; x < 9; x++) {
         if (data[x]->is_set_to(val)) {
             return true;
         }
@@ -40,7 +40,7 @@ bool RowWrapper::contains(const Value val) const
 size_t RowWrapper::possible_places_for(const Value val) const
 {
     size_t result = 0;
-    for (int x = 0; x < 9; x++) {
+    for (size_t x = 0; x < 9; x++) {
         if (data[x]->can_be(val)) {
             result++;
         }
